Load tile layer, through level serialization

+++ src/graphics/tilesetpack.h
@@ -10,8 +10,8 @@ class TileSetPack {
     std::vector<std::tuple<std::shared_ptr<TileSet>, Rectangle*>> tileset_;
 
     int current_tile_id_index_ = 1;
-    uint32_t tile_width_;
-    uint32_t tile_height_;
+    float tile_width_;
+    float tile_height_;
 
    public:
     TileSetPack();
@@ -26,8 +26,8 @@ class TileSetPack {
 
     void Draw(Vector2 const& position, uint16_t tile_id);
     size_t Size();
-    uint32_t GetTileWidth();
-    uint32_t GetTileHeight();
+    float GetTileWidth();
+    float GetTileHeight();
 };
 
 }  // namespace Brutal
